+++ src/kryptos_hash_common.h
@@ -19,7 +19,7 @@ void kryptos_hash_apply_pad_on_u32_block(kryptos_u32_t *input, const size_t inpu
                                          const kryptos_u64_t curr_len, const kryptos_u64_t total_len,
                                          int *paddin2times, const kryptos_u8_t padtok, kryptos_u32_t len_block_offset);
 
-void kryptos_hash_ld_u8buf_as_u32_blocks(kryptos_u8_t *buffer, const int buffer_size,
+void kryptos_hash_ld_u8buf_as_u32_blocks(kryptos_u8_t *buffer, const size_t buffer_size,
                                          kryptos_u32_t *input, const size_t input_nr,
                                          const size_t *block_index_decision_table);
 
@@ -28,7 +28,7 @@ void kryptos_hash_apply_pad_on_u64_block(kryptos_u64_t *input, const size_t inpu
                                          const kryptos_u64_t curr_len, const kryptos_u64_t total_len,
                                          int *paddin2times, const kryptos_u8_t padtok, kryptos_u64_t len_block_offset);
 
-void kryptos_hash_ld_u8buf_as_u64_blocks(kryptos_u8_t *buffer, const int buffer_size,
+void kryptos_hash_ld_u8buf_as_u64_blocks(kryptos_u8_t *buffer, const size_t buffer_size,
                                          kryptos_u64_t *input, const size_t input_nr,
                                          const size_t *block_index_decision_table);
